+++ nc_py_api/talk.py
@@ -1,3 +1,0 @@
-"""Nextcloud API for Talk."""
-
-# /ocs/v2.php/apps/spreed/api/v4