+++ nc_py_api/ex_app/__init__.py
@@ -1,5 +1,5 @@
 """All possible ExApp stuff for NextcloudApp that can be used."""
 from .defs import ApiScope, LogLvl
-from .integration_fastapi import nc_app
+from .integration_fastapi import nc_app, set_handlers
 from .ui.files import UiActionFileInfo, UiFileActionHandlerInfo
 from .uvicorn_fastapi import run_app